requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
openai>=1.0.0
beautifulsoup4>=4.12.0
sentry-sdk>=1.30.0
apscheduler>=3.10.0
python-multipart>=0.0.6
jinja2>=3.1.0
pillow>=10.0.0
weasyprint>=60.0
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.31.0 